날씨 데이터 에러 수정

+++ app/src/main/java/com/seoul/ttarawa/data/remote/response/WeatherResponse.kt
@@ -38,7 +38,7 @@ data class WeatherResponse(
                     @SerializedName("fcstTime")
                     val fcstTime: Int,
                     @SerializedName("fcstValue")
-                    val fcstValue: Int,
+                    val fcstValue: Double,
                     @SerializedName("nx")
                     val nx: Int,
                     @SerializedName("ny")
